requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alembic==1.6.5
alice-blue==1.0.1
alpha-vantage==2.3.1
altgraph==0.17
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==19.11.2
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bidict==0.21.2
binance-aio==0.0.3
binance.py==1.8.3
bleach==3.3.0
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cliff==3.8.0
cmaes==0.8.2
cmd2==2.0.1
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
configparser==5.0.2
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
dateparser==1.0.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.3
dirtyjson==1.0.7
distlib==0.3.2
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
finta==1.3
Flask==2.0.0
flatbuffers==1.12
future==0.18.2
fxcmpy==1.2.9
gast==0.4.0
gevent==21.8.0
google-auth==1.30.2
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
hyperlink==21.0.0
ibapi==9.81.1.post1
idna==2.10
incremental==21.3.0
indicator==1.0.5
interactive-broker-python-web-api==0.1.2
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
Js2Py==0.71
jsonlib-python3==1.6.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiteconnect==3.9.4
kiwisolver==1.3.1
loguru==0.5.3
lxml==4.6.3
macholib==1.14
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==4.7.6
multiprocess==0.70.11.1
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
nsepython==0.0.972
numpy==1.19.5
oandapyV20==0.7.0
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
papertrade==0.1.1
parso==0.8.2
patsy==0.5.1
pbkdf2==1.3
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipwin==0.5.1
platformdirs==2.2.0
playsound==1.2.2
plotly==5.1.0
pox==0.2.9
ppft==1.6.6.3
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protlib==1.4
protobuf==3.17.3
ptyprocess==0.7.0
py2exe==0.9.2.2
py5paisa==0.2.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.2
Pygments==2.9.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyjsparser==2.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyPrind==2.11.3
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pyrsistent==0.18.0
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==3.1.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.5.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
Send2Trash==1.7.1
service-identity==21.1.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smartapi-python==1.2.7
soupsieve==2.2.1
splinter==0.15.0
SQLAlchemy==1.4.18
statsmodels==0.12.2
stevedore==3.3.0
stocknotebridge==1.0.1
stocktrends==0.1.4
tabulate==0.8.9
td-ameritrade-python-api==0.3.5
tdameritrade==0.1.1
telebot==0.0.4
telepot==12.7
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
tweepy==3.10.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ui==0.1.4
ujson==4.0.2
urllib3==1.26.4
virtualenv==20.7.2
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.58.0
websockets==8.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yahoo-finance==1.4.0
yahoofinancials==1.6
yarl==1.6.3
yfinance==0.1.63
zope.event==4.5.0
zope.interface==5.4.0